+++ src/engine/time.c
@@ -43,7 +43,7 @@ uint64_t Time_NowNs(void)
 {
 	struct timespec ts;
 	clock_gettime(CLOCK_MONOTONIC, &ts);
-	return ts.tv_sec * 1000 * 1000 * 1000 + ts.tv_nsec;
+	return (uint64_t)ts.tv_sec * 1000 * 1000 * 1000 + (uint64_t)ts.tv_nsec;
 }
 
 #endif /* _WIN32 */
